Really good progress today, AI is steering really well now and slowes down into bends much better now, also track is orienated correctly now

+++ HotWheels/src/Ai.cpp
@@ -1,10 +1,17 @@
+/**
+ * @class	Ai
+ *
+ * @brief	Article intelligences.
+ *
+ * @author	Ciarán McCann
+ */
+
 #include "stdafx.h"
 #include "Ai.h"
 
-
-Ai::Ai(): mCurrentWayPiont(10), count(0)
+Ai::Ai(): mCurrentWayPiont(2), count(0)
 {
- 
+  
 }
 
 Ai::~Ai(void)
@@ -14,7 +21,12 @@ Ai::~Ai(void)
 
 void Ai::init(Critter::RenderSystem* render, Ogre::SceneManager* mSceneMgr){
 	pCar = Vehicle::createVehicle("camaro",render,  mSceneMgr);
-	pCar->getBody()->setGlobalPosition(Vector3(350.36890, 299.98193, -569.52747));	
+
+	NxOgreUtils::rotate(pCar->getBody(), Vector3(0,1,0), -45);
+	pCar->getBody()->setGlobalPosition(Vector3(340.36890, 299.98193, -569.52747));	
+
+	pCar->setMotorToqureMultplier(2.5);
+	//pCar->setSteeringCoef(0.008f);
 	this->mSceneMgr = mSceneMgr;
 
 }
@@ -40,43 +52,29 @@ void Ai::update(float timeDelta){
 		target.y = 0;
 
 		Vector3 forwardDir =  Ogre::Quaternion(q.w, q.x, q.y, q.z) *  Vector3( 1, 0,1 );
-
 		Vector3 destVector =  target -  Vector3( this->getVechicle()->getBody()->getGlobalPosition().x, 0 , this->getVechicle()->getBody()->getGlobalPosition().z );
-
 		Vector3 pos = Vector3( this->getVechicle()->getBody()->getGlobalPosition().x, this->getVechicle()->getBody()->getGlobalPosition().y , this->getVechicle()->getBody()->getGlobalPosition().z );
 
 
-
-
 		
 
-		mSceneMgr->destroyAllManualObjects();
+		if(Settings::debugMode && false){
 
-		{	
-			ManualObject* manual =  mSceneMgr->createManualObject("manual");
+			mSceneMgr->destroyAllManualObjects();
+			{	
+				ManualObject* manual =  mSceneMgr->createManualObject("manual");
 
+				manual->begin("BeachStones", RenderOperation::OT_LINE_STRIP);
+				manual->colour(0.2,0.0,1.0,1);
+				Vector3 pos = Vector3( this->getVechicle()->getBody()->getGlobalPosition().x, this->getVechicle()->getBody()->getGlobalPosition().y , this->getVechicle()->getBody()->getGlobalPosition().z );
+				pos.y += 2;
+				manual->position(pos);		
+				manual->position(  Vector3( *pTargetPosition ) );
+				manual->end();
 
-			// specify the material (by name) and rendering type
-		
-			manual->begin("BeachStones", RenderOperation::OT_LINE_STRIP);
-			manual->colour(0.2,0.0,1.0,1);
-			// define start and end point
-			Vector3 pos = Vector3( this->getVechicle()->getBody()->getGlobalPosition().x, this->getVechicle()->getBody()->getGlobalPosition().y , this->getVechicle()->getBody()->getGlobalPosition().z );
-			pos.y += 2;
-			manual->position(pos);
-
-
-		
-			manual->position(  Vector3( *pTargetPosition ) );
-
-			// tell Ogre, your definition has finished
-			manual->end();
-
-			mSceneMgr->getRootSceneNode()->createChildSceneNode()->attachObject(manual);
-
+				mSceneMgr->getRootSceneNode()->createChildSceneNode()->attachObject(manual);
+			}
 		}
-
-		
 
 	
 		destVector.y = 0;
@@ -85,34 +83,20 @@ void Ai::update(float timeDelta){
 
 		forwardDir.normalise();
 		destVector.normalise();
+		
+		if(Settings::debugMode && false){
 	
-	
-		{
-
-
+			{
 			ManualObject* manual =  mSceneMgr->createManualObject("manual2");
-
-
-			// specify the material (by name) and rendering type
 			manual->begin("BeachStones", RenderOperation::OT_LINE_STRIP);
-
-			// define start and end point
 			Vector3 pos = Vector3( this->getVechicle()->getBody()->getGlobalPosition().x, this->getVechicle()->getBody()->getGlobalPosition().y , this->getVechicle()->getBody()->getGlobalPosition().z );
 			pos.y += 2;
 			manual->position(pos );
-
-	
-
-
 			forwardDir *= 10.2;
 			manual->position( (pos + forwardDir));
-
-
-			// tell Ogre, your definition has finished
 			manual->end();
-
 			mSceneMgr->getRootSceneNode()->createChildSceneNode()->attachObject(manual);
-
+			}
 		}
 
 
@@ -145,22 +129,25 @@ void Ai::update(float timeDelta){
 
 		Vec3 vTemp = getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,-100));
 		Vector3 v = Vector3(vTemp.x, 0, vTemp.z);
+
+		Vec3 vTemp2 = getVechicle()->getBody()->getLocalPointVelocity(Vec3(10000,0,10000));
+		Vector3 v2 = Vector3(vTemp2.x, 0, vTemp2.z);
+
 		Vector3 wheelNormal = (this->getVechicle()->getWheel(3)->getNode()->getOrientation() * Ogre::Vector3(1,0,0)).normalisedCopy();
 		wheelNormal = wheelNormal * xRotation;
 		
 
 		Utils::log(Ogre::Degree( wheelNormal.angleBetween(v) ).valueDegrees());
-		Utils::drawLine( pos  , pos + (v*v.length()), mSceneMgr );
+
+		//Utils::drawLine( pos  , pos + (v.normalisedCopy()*v.length()), mSceneMgr );
+		//Utils::drawLine( pos  , pos + (v2.normalisedCopy()*v2.length()), mSceneMgr );
 
 			
-		if( int(angle) > 1.0f ){
+		if( int(angle) > 8.0f){
 
-			
-
-			//if(Ogre::Degree( wheelNormal.angleBetween(v) ).valueDegrees() > 80 && Ogre::Degree( wheelNormal.angleBetween(v) ).valueDegrees() < 110 )
 				this->getVechicle()->steerLeft();
 		}
-		else if ( int(angle) < -1.0f) {
+		else if ( int(angle) < -8.0f) {
 
 			Vec3 vTemp = getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,100));
 			Vector3 v = Vector3(vTemp.x, 0, vTemp.z);
@@ -171,58 +158,50 @@ void Ai::update(float timeDelta){
 		}
 
 
-		if( distance > 10 ){
-
-			//if(getVechicle()->getBody()->getLinearVelocity().magnitude() < 6)
+		if( distance > 20 ){
 
 
-			//Utils::log( getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,200,-100)).magnitude() );
+			if( distance < 28 ){
 
-			if( getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,100)).magnitude() < 20
-				&& getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,-100)).magnitude() < 20
-				
+				if( getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,100)).magnitude() < 80
+					&& getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,-100)).magnitude() < 80			
 				)
 				this->getVechicle()->accelerate();
 
+			}else if ( distance < 23 ){
+
+
+					if( getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,100)).magnitude() < 70
+						&& getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,-100)).magnitude() < 70			
+						)
+						this->getVechicle()->accelerate();
+
+			}else{
+
+					if( getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,100)).magnitude() < 100
+					&& getVechicle()->getBody()->getLocalPointVelocity(Vec3(-100,0,-100)).magnitude() < 100	
+				)
+				this->getVechicle()->accelerate();
+			}
+
 		}else{
 
-			//a = -99999;
-			mCurrentWayPiont += 3;
+			mCurrentWayPiont += 5; // Moves AI to next waypoint
 		
-			if(!pTrack->checkIfFinished(mCurrentWayPiont))
-			pTrack->getTrackNodeAt(mCurrentWayPiont)->scale(10,10,10);
+			//if(!pTrack->checkIfFinished(mCurrentWayPiont))
+			//pTrack->getTrackNodeAt(mCurrentWayPiont)->scale(10,10,10);
 
 		}
 
 	}else{
-
+		
+		//AI car has reached the finish line.
 		if(this->getVechicle()->getBody()->getLinearVelocity().magnitude() > 2)
 			this->getVechicle()->deaccelerate();
 
 	}
 }
 
-void Ai::onKeyPress(OIS::Keyboard* pKeyBoard){
-
-	if (pKeyBoard->isKeyDown(OIS::KC_I)){
-		pCar->accelerate();
-	}else if(pKeyBoard->isKeyDown(OIS::KC_K)){
-		pCar->deaccelerate();
-	}
-	
-	if(pKeyBoard->isKeyDown(OIS::KC_J))
-	{
-		pCar->steerRight();
-
-	}else if (pKeyBoard->isKeyDown(OIS::KC_L)){
-		
-
-		pCar->steerLeft();
-
-	}
-
-
-}
 
 Vehicle* Ai::getVechicle(){
 	return pCar;
@@ -230,6 +209,7 @@ Vehicle* Ai::getVechicle(){
 
 void Ai::setTrack(Track * pTrack){
 	this->pTrack = pTrack;
-	if(!Track::trackingEnabled)
-		pTrack->getTrackNodeAt(mCurrentWayPiont)->scale(10,10,10);
+
+	//if(!Track::trackingEnabled)
+		//pTrack->getTrackNodeAt(mCurrentWayPiont)->scale(10,10,10);
 }
